Move entry files into entries directory
